requirements.txt
stix2
python-dateutil
